fix: iterate over keys manually in ProvideMany
Consumers use type assertions to know if it is safe to send us ProvideMany calls, however we would silently drop them on the floor which is uncool.
Fallback to iterating instead.

Something smarter like doing X many concurrent workers might be nice but I feel this needs to go with a bigger DHT refactor and have ProvideMany operations on all router types.